consolidate collection
