+++ EasyReflectometry/experiment/__init__.py
@@ -1,5 +1,5 @@
 from .model import Model
-from .models import Models
+from .model_collection import ModelCollection
 from .resolution_functions import constant_resolution_function
 from .resolution_functions import linear_spline_resolution_function
 
@@ -7,5 +7,5 @@ __all__ = (
     constant_resolution_function,
     linear_spline_resolution_function,
     Model,
-    Models,
+    ModelCollection,
 )
