--- EasyReflectometry/sample/elements/base_element_collection.py
@@ -1,35 +1,0 @@
-from abc import abstractmethod
-
-import yaml
-from easyCore.Objects.Groups import BaseCollection
-
-
-class BaseElementCollection(BaseCollection):
-    def __init__(
-        self,
-        name: str,
-        interface,
-        *args,
-        **kwargs,
-    ):
-        super().__init__(name, *args, **kwargs)
-        self.interface = interface
-
-    @abstractmethod
-    def _dict_repr(self) -> dict[str, str]: ...
-
-    @property
-    def uid(self) -> int:
-        """
-        :return: UID from the borg map
-        """
-        return self._borg.map.convert_id_to_key(self)
-
-    def __repr__(self) -> str:
-        """
-        String representation of the layer.
-
-        :return: a string representation of the layer
-        :rtype: str
-        """
-        return yaml.dump(self._dict_repr, sort_keys=False)
