+++ EasyReflectometry/sample/elements/layers/layer_collection.py
@@ -2,11 +2,11 @@ from __future__ import annotations
 
 __author__ = 'github.com/arm61'
 
-from ..base_element_collection import BaseElementCollection
-from .layer import Layer
+from typing import Optional
 
-# DEFAULT_LAYERS = (Layer(), Layer())
-NR_DEFAULT_LAYERS = 2
+from ...base_element_collection import SIZE_DEFAULT_COLLECTION
+from ...base_element_collection import BaseElementCollection
+from .layer import Layer
 
 
 class LayerCollection(BaseElementCollection):
@@ -15,37 +15,12 @@ class LayerCollection(BaseElementCollection):
 
     def __init__(
         self,
-        *list_layer_like: tuple[Layer],
+        *layers: Optional[list[Layer]],
         name: str = 'EasyLayers',
         interface=None,
         **kwargs,
     ):
-        if not list_layer_like:
-            list_layer_like = tuple([Layer(interface=interface) for _ in range(NR_DEFAULT_LAYERS)])
-        #            layers = DEFAULT_LAYERS
+        if not layers:
+            layers = [Layer(interface=interface) for _ in range(SIZE_DEFAULT_COLLECTION)]
 
-        super().__init__(name, interface, *list_layer_like, **kwargs)
-
-    # Representation
-    @property
-    def _dict_repr(self) -> dict[str, list[dict]]:
-        """
-        A simplified dict representation.
-
-        :return: Simple dictionary
-        """
-        return {self.name: [i._dict_repr for i in self]}
-
-    @classmethod
-    def from_dict(cls, data: dict) -> LayerCollection:
-        """
-        Create a LayerCollection from a dictionary.
-
-        :param data: dictionary of the LayerCollection
-        :return: LayerCollection
-        """
-        layer_collection = super().from_dict(data)
-        # Remove the default layers
-        for i in range(NR_DEFAULT_LAYERS):
-            del layer_collection[0]
-        return layer_collection
+        super().__init__(name, interface, *layers, **kwargs)
